+++ App/public/index.html
@@ -6,28 +6,6 @@
         <link rel="icon" href="../images/logo.png">
     </head>
     <body>
-        <div id="app-container">
-
-        </div>
-
-        <!-- Fontawesome Icons -->
-        <script src="https://kit.fontawesome.com/ff85fda485.js" crossorigin="anonymous"></script>
-        <!-- React -->
-        <script src="https://unpkg.com/babel-standalone@6/babel.min.js"></script>
-        <script src="https://unpkg.com/react@18/umd/react.development.js" crossorigin></script>
-        <script src="https://unpkg.com/react-dom@18/umd/react-dom.development.js" crossorigin></script>
-        <!-- React components -->
-        <script src="app/header.jsx" type="text/babel"></script>
-
-        <script type="text/babel">
-            //Temporary demo script
-            let root = ReactDOM.createRoot(document.getElementById("app-container"));
-            root.render(
-                <div>
-                    <Header exclude="home"/>
-                    <Banner />
-                </div>
-            );
-        </script>
+        <div id="root"></div>
     </body>
 </html>
