+++ App/public/oldindex.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <title>Lads Coffee Co.</title>
+        <link rel="stylesheet" href="style.css">
+        <link rel="icon" href="../images/logo.png">
+    </head>
+    <body>
+        <div id="root"></div>
+
+        </div>
+
+        <!-- Fontawesome Icons -->
+        <script src="https://kit.fontawesome.com/ff85fda485.js" crossorigin="anonymous"></script>
+        <!-- React -->
+        <script src="https://unpkg.com/babel-standalone@6/babel.min.js"></script>
+        <script src="https://unpkg.com/react@18/umd/react.development.js" crossorigin></script>
+        <script src="https://unpkg.com/react-dom@18/umd/react-dom.development.js" crossorigin></script>
+        <!-- React components -->
+        <script src="app/header.jsx" type="text/babel"></script>
+
+        <script type="text/babel">
+            //Temporary demo script
+            let root = ReactDOM.createRoot(document.getElementById("app-container"));
+            root.render(
+                <div>
+                    <Header exclude="home"/>
+                    <Banner />
+                </div>
+            );
+        </script>
+    </body>
+</html>
