--- App/src/about-us.html
@@ -1,84 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <title>About Us</title>
-        <link rel="stylesheet" href="style.css">
-        <link rel="icon" href="../images/logo.png">
-    </head>
-    <body>
-        <nav>
-            <a id="skip-link" tabindex="0" href="#content">Skip to content.</a>
-            <!-- Upper section, with logo and account btns -->
-            <div id="upper-container">
-                <div id="logo-container">
-                    <a href="index.html">
-                        <img src="../images/logo.png" alt="Lads Coffee Co. Logo" id="logo">
-                    </a>
-                </div>
-                <div id="tools">
-                    <a href="search.html">
-                        <i aria-hidden="true" class="fas fa-magnifying-glass" title="Search"></i>
-                        <span class="sr-only">Search</span>
-                    </a>
-                    <a href="account.html">
-                        <i aria-hidden="true" class="fas fa-user" title="My Account"></i>
-                        <span class="sr-only">Account</span>
-                    </a>
-                    <a href="cart.html">
-                        <i aria-hidden="true" class="fas fa-cart-shopping" title="My Cart"></i>
-                        <span class="sr-only">My Cart</span>
-                    </a>
-                </div>
-            </div>
-            <!-- Lower section, with page links -->
-            <div id="page-buttons">
-                <div class="flex-dummy"></div>
-                <a href="about-us.html" class="nav-button">About Us</a>
-                <a href="shop.html" class="nav-button">Shop</a>
-                <a href="subscription.html" class="nav-button">Subscription</a>
-                <a href="map.html" class="nav-button">Locations</a>
-                <div class="flex-dummy"></div>
-            </div>
-        </nav>
-
-
-        <div>
-            <!--First Row-->
-            <section class="aboutus--row">
-                <div class="aboutus--column">
-                    <h1 class="aboutus--h1">Redefining what it means to brew great coffee.</h1>
-                </div>
-                <div class="aboutus--column">
-                    <img src="../images/hotcupwithbeans.jpeg" alt="A cup of coffee" class="aboutus--img">
-                </div>
-            </section>
-
-            <!--Second Row-->
-            <section class="aboutus--row">
-                <div class="aboutus--column" id="aboutus--subscription-wrapper">
-                    <h1 id="aboutus--subscription-title">How does the subscription work?</h1>
-                    <h4 id="aboutus--subscription-body">We know it can be overwhelming to choose your coffee before experiencing that fresh ground smell first.
-                        Our subscription program allows you to try a different coffee every month, or if you are a conniseur
-                        and you know your blend, just select that type at checkout and you'll never run out of 
-                        your favorite beans again!
-                    </h4>
-                </div>
-                
-            </section>
-
-            <!--Third Row-->
-            <section class="aboutus--row">
-                <div class="aboutus--column">
-                    <img src="../images/coffeepeople.jpeg" alt="People drinking coffee and smiling" class="aboutus--img">
-                </div>
-                <div class="aboutus--column">
-                    <h1 class="aboutus--h1">Our Mission</h1>
-                    <h3 class="aboutus--h3">We want to connect our coffee with as many hearts as possible. There are billions of people drinking
-                        filthy, nasty, rotten, bunk coffee beans from some of our competitiors because they're simply clueless. We are 
-                        going to show them what it means to have a good cup of coffee even if it kills us.
-                    </h3>
-                </div>
-            </section>
-        </div>
-    </body>
-</html>
