--- App/src/shop.html
@@ -1,98 +1,0 @@
-﻿<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <title>Lads Coffee Co.</title>
-        <link rel="stylesheet" href="style.css">
-        <link rel="icon" href="../images/logo.png">
-    </head>
-    <body>
-        <nav>
-            <a id="skip-link" tabindex="0" href="#content">Skip to content.</a>
-            <!-- Upper section, with logo and account btns -->
-            <div id="upper-container">
-                <img src="../images/Banner.png" alt="Lads Nav Banner" id="Banner">
-                <div id="tools">
-                    <a href="search.html">
-                        <i aria-hidden="true" class="fas fa-magnifying-glass white-color" title="Search"></i>
-                        <span class="sr-only">Search</span>
-                    </a>
-                    <a href="account.html">
-                        <i aria-hidden="true" class="fas fa-user white-color" title="My Account"></i>
-                        <span class="sr-only">Account</span>
-                    </a>
-                    <a href="cart.html">
-                        <i aria-hidden="true" class="fas fa-cart-shopping white-color" title="My Cart"></i>
-                        <span class="sr-only">My Cart</span>
-                    </a>
-                </div>
-            </div>
-            <!-- Lower section, with page links -->
-            <div id="page-buttons">
-                <div class="flex-dummy"></div>
-                <a href="index.html" class="nav-button">Home</a>
-                <a href="about-us.html" class="nav-button">About Us</a>
-                <a href="subscription.html" class="nav-button">Subscription</a>
-                <a href="map.html" class="nav-button">Locations</a>
-                <div class="flex-dummy"></div>
-            </div>
-            <div>
-                <section class="showcase-container" id="Azetec-Bean">
-                    <div class="item-showcase">
-                        <h1 class="section-header">Azetec Bean</h1>
-                        <img src="../images/Aztec-Bean.png" alt="Azetec Bean coffee" class="item-image" width="500">
-                        <p class="section-text"></p>
-                        <div id="container">
-                            <button>
-                                <span class="add-to-cart">Add To cart</span>
-                            </button>
-                        </div>
-                    </div>
-                    <div id="item-showcase">
-                        <h1 class="section-header">Brazil Bean</h1>
-                        <img src="../images/Brazil-Beans.png" alt="Brazil Bean coffee" class="item-image" width="500">
-                        <p class="section-text"></p>
-                        <div id="container">
-                            <button>
-                                <span class="add-to-cart">Add To cart</span>
-                            </button>
-                        </div>
-                    </div>
-                    <div id="item-showcase">
-                        <h1 class="section-header">Mocha Bag</h1>
-                        <img src="../images/Mocha Bag.png" alt="Mocha Bag" class="item-image" width="500">
-                        <p class="section-text"></p>
-                        <div id="container">
-                            <button>
-                                <span class="add-to-cart">Add To cart</span>
-                            </button>
-                        </div>
-                    </div>
-                </section>
-                <section class="showcase-container" id="Robusta-Bean">
-                    <div class="item-showcase">
-                        <h1 class="section-header">Robusta Bean</h1>
-                        <img src="../images/Robusta-Bean.png" alt="Robusta Bean coffee" class="item-image" width="500">
-                        <p class="section-text"></p>
-                        <div id="container">
-                            <button>
-                                <span class="add-to-cart">Add To cart</span>
-                            </button>
-                        </div>
-                    </div>
-                    <div id="item-showcase">
-                        <h1 class="section-header">Owner's Choice</h1>
-                        <img src="../images/TheClassic-Bean.png" alt="The Classic Bean coffee" class="item-image" width="500">
-                        <p class="section-text"></p>
-                        <div id="container">
-                            <button>
-                                <span class="add-to-cart">Add To cart</span>
-                            </button>
-                        </div>
-                    </div>
-                </section>
-                <section class="showcase-container" id="top-sellers">
-                </section><script src="https://kit.fontawesome.com/ff85fda485.js" crossorigin="anonymous"></script>    
-            </div>
-        </nav>
-    </body>
-</html>
